+++ website/documents/guides/01-getting-started.md
@@ -1,21 +1,6 @@
 ---
 title: Getting Started
 ---
-
-## Installation
-Crank is available on [NPM](https://npmjs.org/@bikeshaving/crank) in the ESModule and CommonJS formats.
-
-```shell
-$ npm install @b9g/crank
-```
-
-```jsx
-import {createElement} from "@b9g/crank";
-import {renderer} from "@b9g/crank/dom";
-
-renderer.render(<div id="hello">Hello world</div>, document.body);
-```
-
 ## Key Examples
 ### A Simple Component
 ```jsx live
@@ -27,101 +12,4 @@ function Greeting({name = "World"}) {
 }
 
 renderer.render(<Greeting />, document.body);
-```
-
-### A Stateful Component
-```jsx live
-import {createElement} from "@b9g/crank";
-import {renderer} from "@b9g/crank/dom.js";
-
-function *Timer() {
-  let seconds = 0;
-  const interval = setInterval(() => {
-    seconds++;
-    this.refresh();
-  }, 1000);
-  try {
-    for ({} of this) {
-      yield <div>Time elapsed {seconds}s</div>;
-    }
-  } finally {
-    clearInterval(interval);
-  }
-}
-
-renderer.render(<Timer />, document.body);
-```
-
-### An Async Component
-```jsx live
-import {createElement} from "@b9g/crank";
-import {renderer} from "@b9g/crank/dom.js";
-
-async function RandomQuote() {
-  const res = await fetch("https://favqs.com/api/qotd");
-  const {quote} = await res.json();
-  return (
-    <figure>
-      <blockquote>{quote.body}</blockquote>
-      <figcaption>
-        - <a href={quote.url} target="_blank">{quote.author}</a>
-      </figcaption>
-    </figure>
-  );
-}
-
-renderer.render(<RandomQuote />, document.body);
-```
-
-### A Loading Component
-```jsx live
-import {createElement} from "@b9g/crank";
-import {renderer} from "@b9g/crank/dom.js";
-
-async function LoadingIndicator() {
-  await new Promise(resolve => setTimeout(resolve, 1000));
-  return <div>Fetching a good boy...</div>;
-}
-
-async function RandomDog({throttle = false}) {
-  const res = await fetch("https://dog.ceo/api/breeds/image/random");
-  const data = await res.json();
-  if (throttle) {
-    await new Promise(resolve => setTimeout(resolve, 2000));
-  }
-
-  return (
-    <img src={data.message} alt="A Random Dog" width="300" />
-  );
-}
-
-async function *RandomDogLoader({throttle}) {
-  for await ({throttle} of this) {
-    yield <LoadingIndicator />;
-    yield <RandomDog throttle={throttle} />;
-  }
-}
-
-function *RandomDogApp() {
-  let throttle = false;
-  this.addEventListener("click", (ev) => {
-    if (ev.target.tagName === "BUTTON") {
-      throttle = !throttle;
-      this.refresh();
-    }
-  });
-
-  for ({} of this) {
-    yield (
-      <>
-        <div>
-          <button>Show me another dog.</button>
-        </div>
-        <RandomDogLoader throttle={throttle} />
-      </>
-    );
-  }
-}
-
-renderer.render(<RandomDogApp />, document.body);
 ```
