+++ website/src/client.css
@@ -247,15 +247,15 @@ code.inline {
 	@media screen and (min-width: 800px) {
 		margin-left: 240px;
 		padding: 2rem 1rem;
-		max-width: 800px;
+		/* max-width: 800px; */
 	}
 
 	@media screen and (min-width: 1100px) {
 		margin-left: 320px;
 		padding: 3rem 2rem;
-		max-width: 1100px;
+		/* max-width: 1100px; */
 		> * {
-			max-width: 800px;
+			/* max-width: 800px; */
 		}
 	}
 }
