--- website/src/components/parts-of-jsx.ts
@@ -1,223 +1,0 @@
-import {jsx} from "@b9g/crank";
-
-export function PartsOfJSX() {
-	return jsx`
-		<svg
-			width="1180"
-			height="268"
-			viewBox="0 0 1180 268"
-			fill="none"
-			xmlns="http://www.w3.org/2000/svg"
-		>
-			<rect x="-5%" y="-5%" width="110%" height="110%" fill="#111b27" />
-			<line
-				x1="762"
-				y1="113.5"
-				x2="1010"
-				y2="113.5"
-				stroke="#c699e3"
-				stroke-width="3"
-			/>
-			<line
-				x1="1008.5"
-				y1="112"
-				x2="1008.5"
-				y2="97"
-				stroke="#c699e3"
-				stroke-width="3"
-			/>
-			<line
-				x1="763.5"
-				y1="97"
-				x2="763.5"
-				y2="112"
-				stroke="#c699e3"
-				stroke-width="3"
-			/>
-			<line
-				x1="805.5"
-				y1="212"
-				x2="805.5"
-				y2="112"
-				stroke="#c699e3"
-				stroke-width="3"
-			/>
-			<line
-				x1="807"
-				y1="210.5"
-				x2="841"
-				y2="210.5"
-				stroke="#c699e3"
-				stroke-width="3"
-			/>
-			<path
-				d="M862.578 217.109C861.807 217.557 861.01 217.891 860.188 218.109C859.375 218.339 858.542 218.453 857.688 218.453C854.979 218.453 852.859 217.641 851.328 216.016C849.807 214.391 849.047 212.141 849.047 209.266C849.047 206.391 849.807 204.141 851.328 202.516C852.859 200.891 854.979 200.078 857.688 200.078C858.531 200.078 859.354 200.188 860.156 200.406C860.958 200.625 861.766 200.964 862.578 201.422V204.438C861.818 203.76 861.052 203.271 860.281 202.969C859.521 202.667 858.656 202.516 857.688 202.516C855.885 202.516 854.5 203.099 853.531 204.266C852.562 205.432 852.078 207.099 852.078 209.266C852.078 211.422 852.562 213.089 853.531 214.266C854.51 215.432 855.896 216.016 857.688 216.016C858.688 216.016 859.583 215.865 860.375 215.562C861.167 215.25 861.901 214.771 862.578 214.125V217.109ZM881.703 207.156V218H878.812V207.156C878.812 205.583 878.536 204.427 877.984 203.688C877.432 202.948 876.568 202.578 875.391 202.578C874.047 202.578 873.01 203.057 872.281 204.016C871.562 204.964 871.203 206.328 871.203 208.109V218H868.328V193.688H871.203V203.125C871.714 202.125 872.406 201.37 873.281 200.859C874.156 200.339 875.193 200.078 876.391 200.078C878.172 200.078 879.5 200.667 880.375 201.844C881.26 203.01 881.703 204.781 881.703 207.156ZM889.5 200.562H896.859V215.766H902.562V218H888.281V215.766H893.984V202.797H889.5V200.562ZM893.984 193.766H896.859V197.391H893.984V193.766ZM915.078 211.672C915.078 212.964 915.312 213.938 915.781 214.594C916.26 215.25 916.964 215.578 917.891 215.578H921.25V218H917.609C915.891 218 914.557 217.453 913.609 216.359C912.672 215.255 912.203 213.693 912.203 211.672V195.844H907.594V193.594H915.078V211.672ZM936.531 202.734V193.688H939.406V218H936.531V215.797C936.052 216.661 935.411 217.323 934.609 217.781C933.818 218.229 932.901 218.453 931.859 218.453C929.745 218.453 928.078 217.635 926.859 216C925.651 214.354 925.047 212.089 925.047 209.203C925.047 206.359 925.656 204.13 926.875 202.516C928.094 200.891 929.755 200.078 931.859 200.078C932.911 200.078 933.839 200.307 934.641 200.766C935.443 201.214 936.073 201.87 936.531 202.734ZM928.078 209.266C928.078 211.495 928.432 213.177 929.141 214.312C929.849 215.448 930.896 216.016 932.281 216.016C933.667 216.016 934.719 215.443 935.438 214.297C936.167 213.151 936.531 211.474 936.531 209.266C936.531 207.047 936.167 205.37 935.438 204.234C934.719 203.089 933.667 202.516 932.281 202.516C930.896 202.516 929.849 203.083 929.141 204.219C928.432 205.354 928.078 207.036 928.078 209.266ZM960.453 204.109C959.839 203.63 959.214 203.281 958.578 203.062C957.943 202.844 957.245 202.734 956.484 202.734C954.693 202.734 953.323 203.297 952.375 204.422C951.427 205.547 950.953 207.172 950.953 209.297V218H948.062V200.5H950.953V203.922C951.432 202.682 952.167 201.734 953.156 201.078C954.156 200.411 955.339 200.078 956.703 200.078C957.411 200.078 958.073 200.167 958.688 200.344C959.302 200.521 959.891 200.797 960.453 201.172V204.109ZM979.188 207.812V209.219H966.734V209.312C966.651 211.698 967.104 213.411 968.094 214.453C969.094 215.495 970.5 216.016 972.312 216.016C973.229 216.016 974.188 215.87 975.188 215.578C976.188 215.286 977.255 214.844 978.391 214.25V217.109C977.297 217.557 976.24 217.891 975.219 218.109C974.208 218.339 973.229 218.453 972.281 218.453C969.562 218.453 967.438 217.641 965.906 216.016C964.375 214.38 963.609 212.13 963.609 209.266C963.609 206.474 964.359 204.245 965.859 202.578C967.359 200.911 969.359 200.078 971.859 200.078C974.089 200.078 975.844 200.833 977.125 202.344C978.417 203.854 979.104 205.677 979.188 207.812ZM976.312 206.969C976.188 205.76 975.745 204.719 974.984 203.844C974.234 202.958 973.151 202.516 971.734 202.516C970.349 202.516 969.208 202.974 968.312 203.891C967.417 204.807 966.927 205.839 966.844 206.984L976.312 206.969ZM997.391 207.156V218H994.5V207.156C994.5 205.583 994.224 204.427 993.672 203.688C993.12 202.948 992.255 202.578 991.078 202.578C989.734 202.578 988.698 203.057 987.969 204.016C987.25 204.964 986.891 206.328 986.891 208.109V218H984.016V200.5H986.891V203.125C987.401 202.125 988.094 201.37 988.969 200.859C989.844 200.339 990.88 200.078 992.078 200.078C993.859 200.078 995.188 200.667 996.062 201.844C996.948 203.01 997.391 204.781 997.391 207.156Z"
-				fill="#c699e3"
-			/>
-			<line
-				x1="339"
-				y1="113.5"
-				x2="743"
-				y2="113.5"
-				stroke="#e9ae7e"
-				stroke-width="3"
-			/>
-			<line
-				x1="741.5"
-				y1="112"
-				x2="741.5"
-				y2="97"
-				stroke="#e9ae7e"
-				stroke-width="3"
-			/>
-			<line
-				x1="340.5"
-				y1="97"
-				x2="340.5"
-				y2="112"
-				stroke="#e9ae7e"
-				stroke-width="3"
-			/>
-			<line
-				x1="446.5"
-				y1="212"
-				x2="446.5"
-				y2="112"
-				stroke="#e9ae7e"
-				stroke-width="3"
-			/>
-			<line
-				x1="448"
-				y1="210.5"
-				x2="482"
-				y2="210.5"
-				stroke="#e9ae7e"
-				stroke-width="3"
-			/>
-			<path
-				d="M492.859 215.797V224.656H489.969V200.5H492.859V202.734C493.339 201.87 493.974 201.214 494.766 200.766C495.568 200.307 496.49 200.078 497.531 200.078C499.646 200.078 501.302 200.896 502.5 202.531C503.708 204.167 504.312 206.432 504.312 209.328C504.312 212.172 503.708 214.406 502.5 216.031C501.292 217.646 499.635 218.453 497.531 218.453C496.469 218.453 495.536 218.229 494.734 217.781C493.943 217.323 493.318 216.661 492.859 215.797ZM501.297 209.266C501.297 207.036 500.943 205.354 500.234 204.219C499.536 203.083 498.495 202.516 497.109 202.516C495.714 202.516 494.656 203.089 493.938 204.234C493.219 205.37 492.859 207.047 492.859 209.266C492.859 211.474 493.219 213.151 493.938 214.297C494.656 215.443 495.714 216.016 497.109 216.016C498.495 216.016 499.536 215.448 500.234 214.312C500.943 213.177 501.297 211.495 501.297 209.266ZM524.328 204.109C523.714 203.63 523.089 203.281 522.453 203.062C521.818 202.844 521.12 202.734 520.359 202.734C518.568 202.734 517.198 203.297 516.25 204.422C515.302 205.547 514.828 207.172 514.828 209.297V218H511.938V200.5H514.828V203.922C515.307 202.682 516.042 201.734 517.031 201.078C518.031 200.411 519.214 200.078 520.578 200.078C521.286 200.078 521.948 200.167 522.562 200.344C523.177 200.521 523.766 200.797 524.328 201.172V204.109ZM535.188 202.516C533.729 202.516 532.625 203.083 531.875 204.219C531.125 205.354 530.75 207.036 530.75 209.266C530.75 211.484 531.125 213.167 531.875 214.312C532.625 215.448 533.729 216.016 535.188 216.016C536.656 216.016 537.766 215.448 538.516 214.312C539.266 213.167 539.641 211.484 539.641 209.266C539.641 207.036 539.266 205.354 538.516 204.219C537.766 203.083 536.656 202.516 535.188 202.516ZM535.188 200.078C537.615 200.078 539.469 200.865 540.75 202.438C542.042 204.01 542.688 206.286 542.688 209.266C542.688 212.255 542.047 214.536 540.766 216.109C539.484 217.672 537.625 218.453 535.188 218.453C532.76 218.453 530.906 217.672 529.625 216.109C528.344 214.536 527.703 212.255 527.703 209.266C527.703 206.286 528.344 204.01 529.625 202.438C530.906 200.865 532.76 200.078 535.188 200.078ZM550.703 215.797V224.656H547.812V200.5H550.703V202.734C551.182 201.87 551.818 201.214 552.609 200.766C553.411 200.307 554.333 200.078 555.375 200.078C557.49 200.078 559.146 200.896 560.344 202.531C561.552 204.167 562.156 206.432 562.156 209.328C562.156 212.172 561.552 214.406 560.344 216.031C559.135 217.646 557.479 218.453 555.375 218.453C554.312 218.453 553.38 218.229 552.578 217.781C551.786 217.323 551.161 216.661 550.703 215.797ZM559.141 209.266C559.141 207.036 558.786 205.354 558.078 204.219C557.38 203.083 556.339 202.516 554.953 202.516C553.557 202.516 552.5 203.089 551.781 204.234C551.062 205.37 550.703 207.047 550.703 209.266C550.703 211.474 551.062 213.151 551.781 214.297C552.5 215.443 553.557 216.016 554.953 216.016C556.339 216.016 557.38 215.448 558.078 214.312C558.786 213.177 559.141 211.495 559.141 209.266ZM579.328 201.109V203.922C578.505 203.443 577.677 203.083 576.844 202.844C576.01 202.604 575.161 202.484 574.297 202.484C572.995 202.484 572.021 202.698 571.375 203.125C570.74 203.542 570.422 204.182 570.422 205.047C570.422 205.828 570.661 206.411 571.141 206.797C571.62 207.182 572.812 207.557 574.719 207.922L575.875 208.141C577.302 208.411 578.38 208.953 579.109 209.766C579.849 210.578 580.219 211.635 580.219 212.938C580.219 214.667 579.604 216.021 578.375 217C577.146 217.969 575.438 218.453 573.25 218.453C572.385 218.453 571.479 218.359 570.531 218.172C569.583 217.995 568.557 217.724 567.453 217.359V214.391C568.526 214.943 569.552 215.359 570.531 215.641C571.51 215.911 572.438 216.047 573.312 216.047C574.583 216.047 575.568 215.792 576.266 215.281C576.964 214.76 577.312 214.036 577.312 213.109C577.312 211.776 576.036 210.854 573.484 210.344L573.359 210.312L572.281 210.094C570.625 209.771 569.417 209.229 568.656 208.469C567.896 207.698 567.516 206.651 567.516 205.328C567.516 203.651 568.083 202.359 569.219 201.453C570.354 200.536 571.974 200.078 574.078 200.078C575.016 200.078 575.917 200.167 576.781 200.344C577.646 200.51 578.495 200.766 579.328 201.109Z"
-				fill="#e9ae7e"
-			/>
-			<line
-				x1="281"
-				y1="113.5"
-				x2="321"
-				y2="113.5"
-				stroke="#6cb8e6"
-				stroke-width="3"
-			/>
-			<line
-				x1="319.5"
-				y1="112"
-				x2="319.5"
-				y2="97"
-				stroke="#6cb8e6"
-				stroke-width="3"
-			/>
-			<line
-				x1="282.5"
-				y1="97"
-				x2="282.5"
-				y2="112"
-				stroke="#6cb8e6"
-				stroke-width="3"
-			/>
-			<line
-				x1="1053"
-				y1="113.5"
-				x2="1092"
-				y2="113.5"
-				stroke="#6cb8e6"
-				stroke-width="3"
-			/>
-			<line
-				x1="1090.5"
-				y1="112"
-				x2="1090.5"
-				y2="97"
-				stroke="#6cb8e6"
-				stroke-width="3"
-			/>
-			<line
-				x1="1054.5"
-				y1="97"
-				x2="1054.5"
-				y2="112"
-				stroke="#6cb8e6"
-				stroke-width="3"
-			/>
-			<line
-				x1="815"
-				y1="162.5"
-				x2="1071"
-				y2="162.5"
-				stroke="#6cb8e6"
-				stroke-width="3"
-			/>
-			<line
-				x1="300.5"
-				y1="114"
-				x2="300.5"
-				y2="164"
-				stroke="#6cb8e6"
-				stroke-width="3"
-			/>
-			<line
-				x1="302"
-				y1="162.5"
-				x2="336"
-				y2="162.5"
-				stroke="#6cb8e6"
-				stroke-width="3"
-			/>
-			<line
-				x1="1072.5"
-				y1="112"
-				x2="1072.5"
-				y2="164"
-				stroke="#6cb8e6"
-				stroke-width="3"
-			/>
-			<line
-				x1="456"
-				y1="162.5"
-				x2="796"
-				y2="162.5"
-				stroke="#6cb8e6"
-				stroke-width="3"
-			/>
-			<line
-				x1="405"
-				y1="162.5"
-				x2="437"
-				y2="162.5"
-				stroke="#6cb8e6"
-				stroke-width="3"
-			/>
-			<path
-				d="M351.594 152.531V157.5H358.125V159.734H351.594V169.234C351.594 170.526 351.839 171.427 352.328 171.938C352.818 172.448 353.672 172.703 354.891 172.703H358.125V175H354.609C352.453 175 350.932 174.568 350.047 173.703C349.161 172.839 348.719 171.349 348.719 169.234V159.734H344.047V157.5H348.719V152.531H351.594ZM372.25 166.203H371.297C369.62 166.203 368.354 166.5 367.5 167.094C366.656 167.677 366.234 168.552 366.234 169.719C366.234 170.771 366.552 171.589 367.188 172.172C367.823 172.755 368.703 173.047 369.828 173.047C371.411 173.047 372.656 172.5 373.562 171.406C374.469 170.302 374.927 168.781 374.938 166.844V166.203H372.25ZM377.828 165.016V175H374.938V172.406C374.323 173.448 373.547 174.219 372.609 174.719C371.682 175.208 370.552 175.453 369.219 175.453C367.438 175.453 366.016 174.953 364.953 173.953C363.891 172.943 363.359 171.594 363.359 169.906C363.359 167.958 364.01 166.479 365.312 165.469C366.625 164.458 368.547 163.953 371.078 163.953H374.938V163.5C374.927 162.104 374.573 161.094 373.875 160.469C373.177 159.833 372.062 159.516 370.531 159.516C369.552 159.516 368.562 159.656 367.562 159.938C366.562 160.219 365.589 160.63 364.641 161.172V158.297C365.703 157.891 366.719 157.589 367.688 157.391C368.667 157.182 369.615 157.078 370.531 157.078C371.979 157.078 373.214 157.292 374.234 157.719C375.266 158.146 376.099 158.786 376.734 159.641C377.13 160.161 377.411 160.807 377.578 161.578C377.745 162.339 377.828 163.484 377.828 165.016ZM393.969 166.109C393.969 163.953 393.615 162.318 392.906 161.203C392.208 160.078 391.188 159.516 389.844 159.516C388.438 159.516 387.365 160.078 386.625 161.203C385.885 162.318 385.516 163.953 385.516 166.109C385.516 168.266 385.885 169.911 386.625 171.047C387.375 172.172 388.458 172.734 389.875 172.734C391.198 172.734 392.208 172.167 392.906 171.031C393.615 169.896 393.969 168.255 393.969 166.109ZM396.844 173.875C396.844 176.5 396.224 178.49 394.984 179.844C393.745 181.198 391.922 181.875 389.516 181.875C388.724 181.875 387.896 181.802 387.031 181.656C386.167 181.51 385.302 181.297 384.438 181.016V178.172C385.458 178.651 386.385 179.005 387.219 179.234C388.052 179.464 388.818 179.578 389.516 179.578C391.068 179.578 392.198 179.156 392.906 178.312C393.615 177.469 393.969 176.13 393.969 174.297V174.172V172.219C393.51 173.198 392.885 173.927 392.094 174.406C391.302 174.885 390.339 175.125 389.203 175.125C387.161 175.125 385.531 174.307 384.312 172.672C383.094 171.036 382.484 168.849 382.484 166.109C382.484 163.359 383.094 161.167 384.312 159.531C385.531 157.896 387.161 157.078 389.203 157.078C390.328 157.078 391.281 157.302 392.062 157.75C392.844 158.198 393.479 158.891 393.969 159.828V157.562H396.844V173.875Z"
-				fill="#6cb8e6"
-			/>
-			<path
-				d="M66.5781 79.1094C65.8073 79.5573 65.0104 79.8906 64.1875 80.1094C63.375 80.3385 62.5417 80.4531 61.6875 80.4531C58.9792 80.4531 56.8594 79.6406 55.3281 78.0156C53.8073 76.3906 53.0469 74.1406 53.0469 71.2656C53.0469 68.3906 53.8073 66.1406 55.3281 64.5156C56.8594 62.8906 58.9792 62.0781 61.6875 62.0781C62.5312 62.0781 63.3542 62.1875 64.1562 62.4062C64.9583 62.625 65.7656 62.9635 66.5781 63.4219V66.4375C65.8177 65.7604 65.0521 65.2708 64.2812 64.9688C63.5208 64.6667 62.6562 64.5156 61.6875 64.5156C59.8854 64.5156 58.5 65.099 57.5312 66.2656C56.5625 67.4323 56.0781 69.099 56.0781 71.2656C56.0781 73.4219 56.5625 75.0885 57.5312 76.2656C58.5104 77.4323 59.8958 78.0156 61.6875 78.0156C62.6875 78.0156 63.5833 77.8646 64.375 77.5625C65.1667 77.25 65.901 76.7708 66.5781 76.125V79.1094ZM78.9062 64.5156C77.4479 64.5156 76.3438 65.0833 75.5938 66.2188C74.8438 67.3542 74.4688 69.0365 74.4688 71.2656C74.4688 73.4844 74.8438 75.1667 75.5938 76.3125C76.3438 77.4479 77.4479 78.0156 78.9062 78.0156C80.375 78.0156 81.4844 77.4479 82.2344 76.3125C82.9844 75.1667 83.3594 73.4844 83.3594 71.2656C83.3594 69.0365 82.9844 67.3542 82.2344 66.2188C81.4844 65.0833 80.375 64.5156 78.9062 64.5156ZM78.9062 62.0781C81.3333 62.0781 83.1875 62.8646 84.4688 64.4375C85.7604 66.0104 86.4062 68.2865 86.4062 71.2656C86.4062 74.2552 85.7656 76.5365 84.4844 78.1094C83.2031 79.6719 81.3438 80.4531 78.9062 80.4531C76.4792 80.4531 74.625 79.6719 73.3438 78.1094C72.0625 76.5365 71.4219 74.2552 71.4219 71.2656C71.4219 68.2865 72.0625 66.0104 73.3438 64.4375C74.625 62.8646 76.4792 62.0781 78.9062 62.0781ZM104.984 69.1562V80H102.094V69.1562C102.094 67.5833 101.818 66.4271 101.266 65.6875C100.714 64.9479 99.849 64.5781 98.6719 64.5781C97.3281 64.5781 96.2917 65.0573 95.5625 66.0156C94.8438 66.9635 94.4844 68.3281 94.4844 70.1094V80H91.6094V62.5H94.4844V65.125C94.9948 64.125 95.6875 63.3698 96.5625 62.8594C97.4375 62.3385 98.474 62.0781 99.6719 62.0781C101.453 62.0781 102.781 62.6667 103.656 63.8438C104.542 65.0104 104.984 66.7812 104.984 69.1562ZM123.047 63.1094V65.9219C122.224 65.4427 121.396 65.0833 120.562 64.8438C119.729 64.6042 118.88 64.4844 118.016 64.4844C116.714 64.4844 115.74 64.6979 115.094 65.125C114.458 65.5417 114.141 66.1823 114.141 67.0469C114.141 67.8281 114.38 68.4115 114.859 68.7969C115.339 69.1823 116.531 69.5573 118.438 69.9219L119.594 70.1406C121.021 70.4115 122.099 70.9531 122.828 71.7656C123.568 72.5781 123.938 73.6354 123.938 74.9375C123.938 76.6667 123.323 78.0208 122.094 79C120.865 79.9688 119.156 80.4531 116.969 80.4531C116.104 80.4531 115.198 80.3594 114.25 80.1719C113.302 79.9948 112.276 79.724 111.172 79.3594V76.3906C112.245 76.9427 113.271 77.3594 114.25 77.6406C115.229 77.9115 116.156 78.0469 117.031 78.0469C118.302 78.0469 119.286 77.7917 119.984 77.2812C120.682 76.7604 121.031 76.0365 121.031 75.1094C121.031 73.776 119.755 72.8542 117.203 72.3438L117.078 72.3125L116 72.0938C114.344 71.7708 113.135 71.2292 112.375 70.4688C111.615 69.6979 111.234 68.651 111.234 67.3281C111.234 65.651 111.802 64.3594 112.938 63.4531C114.073 62.5365 115.693 62.0781 117.797 62.0781C118.734 62.0781 119.635 62.1667 120.5 62.3438C121.365 62.5104 122.214 62.7656 123.047 63.1094ZM136.719 57.5312V62.5H143.25V64.7344H136.719V74.2344C136.719 75.526 136.964 76.4271 137.453 76.9375C137.943 77.4479 138.797 77.7031 140.016 77.7031H143.25V80H139.734C137.578 80 136.057 79.5677 135.172 78.7031C134.286 77.8385 133.844 76.349 133.844 74.2344V64.7344H129.172V62.5H133.844V57.5312H136.719ZM183.188 69.8125V71.2188H170.734V71.3125C170.651 73.6979 171.104 75.4115 172.094 76.4531C173.094 77.4948 174.5 78.0156 176.312 78.0156C177.229 78.0156 178.188 77.8698 179.188 77.5781C180.188 77.2865 181.255 76.8438 182.391 76.25V79.1094C181.297 79.5573 180.24 79.8906 179.219 80.1094C178.208 80.3385 177.229 80.4531 176.281 80.4531C173.562 80.4531 171.438 79.6406 169.906 78.0156C168.375 76.3802 167.609 74.1302 167.609 71.2656C167.609 68.474 168.359 66.2448 169.859 64.5781C171.359 62.9115 173.359 62.0781 175.859 62.0781C178.089 62.0781 179.844 62.8333 181.125 64.3438C182.417 65.8542 183.104 67.6771 183.188 69.8125ZM180.312 68.9688C180.188 67.7604 179.745 66.7188 178.984 65.8438C178.234 64.9583 177.151 64.5156 175.734 64.5156C174.349 64.5156 173.208 64.974 172.312 65.8906C171.417 66.8073 170.927 67.8385 170.844 68.9844L180.312 68.9688ZM196.203 73.6719C196.203 74.9635 196.438 75.9375 196.906 76.5938C197.385 77.25 198.089 77.5781 199.016 77.5781H202.375V80H198.734C197.016 80 195.682 79.4531 194.734 78.3594C193.797 77.2552 193.328 75.6927 193.328 73.6719V57.8438H188.719V55.5938H196.203V73.6719ZM224.906 72.1875H241.422V74.8281H224.906V72.1875ZM224.906 64.8281H241.422V67.4688H224.906V64.8281ZM1117.38 74.3594H1122.25V78.5L1119.17 84.4688H1115.83L1117.38 78.5V74.3594ZM1117.17 62.5156H1122.05V68.1562H1117.17V62.5156Z"
-				fill="#6B7990"
-			/>
-			<path
-				d="M279.984 64.9844L266.703 69.9844L279.984 74.9375V77.7969L263.469 71.2656V68.6719L279.984 62.1406V64.9844ZM745.5 64.9844V62.1406L762.016 68.6719V71.2656L745.5 77.7969V74.9375L758.781 69.9844L745.5 64.9844ZM1031.95 64.9844L1018.67 69.9844L1031.95 74.9375V77.7969L1015.44 71.2656V68.6719L1031.95 62.1406V64.9844ZM1047.23 56.6719H1050.2L1037.92 82.9688H1034.94L1047.23 56.6719ZM1092.56 64.9844V62.1406L1109.08 68.6719V71.2656L1092.56 77.7969V74.9375L1105.84 69.9844L1092.56 64.9844Z"
-				fill="#6699CC"
-			/>
-			<path
-				d="M297.797 69.1562V80H294.906V69.1562C294.906 67.5833 294.63 66.4271 294.078 65.6875C293.526 64.9479 292.661 64.5781 291.484 64.5781C290.141 64.5781 289.104 65.0573 288.375 66.0156C287.656 66.9635 287.297 68.3281 287.297 70.1094V80H284.422V55.6875H287.297V65.125C287.807 64.125 288.5 63.3698 289.375 62.8594C290.25 62.3385 291.286 62.0781 292.484 62.0781C294.266 62.0781 295.594 62.6667 296.469 63.8438C297.354 65.0104 297.797 66.7812 297.797 69.1562ZM304.875 77.3594H309.781V59.7344L303.906 62.5625V59.5312L309.75 56.75H312.906V77.3594H317.75V80H304.875V77.3594ZM1069.05 69.1562V80H1066.16V69.1562C1066.16 67.5833 1065.88 66.4271 1065.33 65.6875C1064.78 64.9479 1063.91 64.5781 1062.73 64.5781C1061.39 64.5781 1060.35 65.0573 1059.62 66.0156C1058.91 66.9635 1058.55 68.3281 1058.55 70.1094V80H1055.67V55.6875H1058.55V65.125C1059.06 64.125 1059.75 63.3698 1060.62 62.8594C1061.5 62.3385 1062.54 62.0781 1063.73 62.0781C1065.52 62.0781 1066.84 62.6667 1067.72 63.8438C1068.6 65.0104 1069.05 66.7812 1069.05 69.1562ZM1076.12 77.3594H1081.03V59.7344L1075.16 62.5625V59.5312L1081 56.75H1084.16V77.3594H1089V80H1076.12V77.3594Z"
-				fill="#6cb8e6"
-			/>
-			<path
-				d="M344.156 62.5625H351.516V77.7656H357.219V80H342.938V77.7656H348.641V64.7969H344.156V62.5625ZM348.641 55.7656H351.516V59.3906H348.641V55.7656ZM371.906 64.7344V55.6875H374.781V80H371.906V77.7969C371.427 78.6615 370.786 79.3229 369.984 79.7812C369.193 80.2292 368.276 80.4531 367.234 80.4531C365.12 80.4531 363.453 79.6354 362.234 78C361.026 76.3542 360.422 74.0885 360.422 71.2031C360.422 68.3594 361.031 66.1302 362.25 64.5156C363.469 62.8906 365.13 62.0781 367.234 62.0781C368.286 62.0781 369.214 62.3073 370.016 62.7656C370.818 63.2135 371.448 63.8698 371.906 64.7344ZM363.453 71.2656C363.453 73.4948 363.807 75.1771 364.516 76.3125C365.224 77.4479 366.271 78.0156 367.656 78.0156C369.042 78.0156 370.094 77.4427 370.812 76.2969C371.542 75.151 371.906 73.474 371.906 71.2656C371.906 69.0469 371.542 67.3698 370.812 66.2344C370.094 65.0885 369.042 64.5156 367.656 64.5156C366.271 64.5156 365.224 65.0833 364.516 66.2188C363.807 67.3542 363.453 69.0365 363.453 71.2656ZM379.156 72.1875H395.672V74.8281H379.156V72.1875ZM379.156 64.8281H395.672V67.4688H379.156V64.8281ZM413.188 80.7969V83.0469H412.188C409.594 83.0469 407.854 82.6615 406.969 81.8906C406.094 81.1198 405.656 79.5833 405.656 77.2812V74.1562C405.656 73.0208 405.333 72.0417 404.688 71.2188C404.052 70.3958 403.062 69.9844 401.719 69.9844H399.812V67.75H401.719C403.188 67.75 404.208 67.401 404.781 66.7031C405.365 66.0052 405.656 64.9792 405.656 63.625V59.5625C405.656 57.2604 406.094 55.7292 406.969 54.9688C407.854 54.1979 409.594 53.8125 412.188 53.8125H413.188V56.0469H412.094C410.635 56.0469 409.682 56.276 409.234 56.7344C408.797 57.1823 408.578 58.1458 408.578 59.625V63.8125C408.578 65.6458 408.219 66.8802 407.5 67.5156C406.781 68.1406 405.38 68.5938 403.297 68.875C405.318 69.1146 406.703 69.5781 407.453 70.2656C408.203 70.9427 408.578 72.1719 408.578 73.9531V77.2188C408.578 78.7083 408.797 79.6771 409.234 80.125C409.682 80.5729 410.635 80.7969 412.094 80.7969H413.188ZM421.281 62.5625H428.641V77.7656H434.344V80H420.062V77.7656H425.766V64.7969H421.281V62.5625ZM425.766 55.7656H428.641V59.3906H425.766V55.7656ZM449.031 64.7344V55.6875H451.906V80H449.031V77.7969C448.552 78.6615 447.911 79.3229 447.109 79.7812C446.318 80.2292 445.401 80.4531 444.359 80.4531C442.245 80.4531 440.578 79.6354 439.359 78C438.151 76.3542 437.547 74.0885 437.547 71.2031C437.547 68.3594 438.156 66.1302 439.375 64.5156C440.594 62.8906 442.255 62.0781 444.359 62.0781C445.411 62.0781 446.339 62.3073 447.141 62.7656C447.943 63.2135 448.573 63.8698 449.031 64.7344ZM440.578 71.2656C440.578 73.4948 440.932 75.1771 441.641 76.3125C442.349 77.4479 443.396 78.0156 444.781 78.0156C446.167 78.0156 447.219 77.4427 447.938 76.2969C448.667 75.151 449.031 73.474 449.031 71.2656C449.031 69.0469 448.667 67.3698 447.938 66.2344C447.219 65.0885 446.167 64.5156 444.781 64.5156C443.396 64.5156 442.349 65.0833 441.641 66.2188C440.932 67.3542 440.578 69.0365 440.578 71.2656ZM459.422 80.7969C460.88 80.7969 461.828 80.5729 462.266 80.125C462.714 79.6771 462.938 78.7083 462.938 77.2188V73.9531C462.938 72.1719 463.312 70.9427 464.062 70.2656C464.812 69.5781 466.198 69.1146 468.219 68.875C466.135 68.5938 464.734 68.1406 464.016 67.5156C463.297 66.8802 462.938 65.6458 462.938 63.8125V59.625C462.938 58.1458 462.714 57.1823 462.266 56.7344C461.828 56.276 460.88 56.0469 459.422 56.0469H458.328V53.8125H459.328C461.922 53.8125 463.656 54.1979 464.531 54.9688C465.417 55.7292 465.859 57.2604 465.859 59.5625V63.625C465.859 64.9792 466.146 66.0052 466.719 66.7031C467.302 67.401 468.328 67.75 469.797 67.75H471.703V69.9844H469.797C468.453 69.9844 467.458 70.3958 466.812 71.2188C466.177 72.0417 465.859 73.0208 465.859 74.1562V77.2812C465.859 79.5833 465.417 81.1198 464.531 81.8906C463.656 82.6615 461.922 83.0469 459.328 83.0469H458.328V80.7969H459.422ZM510.047 79.1094C509.276 79.5573 508.479 79.8906 507.656 80.1094C506.844 80.3385 506.01 80.4531 505.156 80.4531C502.448 80.4531 500.328 79.6406 498.797 78.0156C497.276 76.3906 496.516 74.1406 496.516 71.2656C496.516 68.3906 497.276 66.1406 498.797 64.5156C500.328 62.8906 502.448 62.0781 505.156 62.0781C506 62.0781 506.823 62.1875 507.625 62.4062C508.427 62.625 509.234 62.9635 510.047 63.4219V66.4375C509.286 65.7604 508.521 65.2708 507.75 64.9688C506.99 64.6667 506.125 64.5156 505.156 64.5156C503.354 64.5156 501.969 65.099 501 66.2656C500.031 67.4323 499.547 69.099 499.547 71.2656C499.547 73.4219 500.031 75.0885 501 76.2656C501.979 77.4323 503.365 78.0156 505.156 78.0156C506.156 78.0156 507.052 77.8646 507.844 77.5625C508.635 77.25 509.37 76.7708 510.047 76.125V79.1094ZM523.984 73.6719C523.984 74.9635 524.219 75.9375 524.688 76.5938C525.167 77.25 525.87 77.5781 526.797 77.5781H530.156V80H526.516C524.797 80 523.464 79.4531 522.516 78.3594C521.578 77.2552 521.109 75.6927 521.109 73.6719V57.8438H516.5V55.5938H523.984V73.6719ZM543 71.2031H542.047C540.37 71.2031 539.104 71.5 538.25 72.0938C537.406 72.6771 536.984 73.5521 536.984 74.7188C536.984 75.7708 537.302 76.5885 537.938 77.1719C538.573 77.7552 539.453 78.0469 540.578 78.0469C542.161 78.0469 543.406 77.5 544.312 76.4062C545.219 75.3021 545.677 73.7812 545.688 71.8438V71.2031H543ZM548.578 70.0156V80H545.688V77.4062C545.073 78.4479 544.297 79.2188 543.359 79.7188C542.432 80.2083 541.302 80.4531 539.969 80.4531C538.188 80.4531 536.766 79.9531 535.703 78.9531C534.641 77.9427 534.109 76.5938 534.109 74.9062C534.109 72.9583 534.76 71.4792 536.062 70.4688C537.375 69.4583 539.297 68.9531 541.828 68.9531H545.688V68.5C545.677 67.1042 545.323 66.0938 544.625 65.4688C543.927 64.8333 542.812 64.5156 541.281 64.5156C540.302 64.5156 539.312 64.6562 538.312 64.9375C537.312 65.2188 536.339 65.6302 535.391 66.1719V63.2969C536.453 62.8906 537.469 62.5885 538.438 62.3906C539.417 62.1823 540.365 62.0781 541.281 62.0781C542.729 62.0781 543.964 62.2917 544.984 62.7188C546.016 63.1458 546.849 63.7865 547.484 64.6406C547.88 65.1615 548.161 65.8073 548.328 66.5781C548.495 67.3385 548.578 68.4844 548.578 70.0156ZM566.516 63.1094V65.9219C565.693 65.4427 564.865 65.0833 564.031 64.8438C563.198 64.6042 562.349 64.4844 561.484 64.4844C560.182 64.4844 559.208 64.6979 558.562 65.125C557.927 65.5417 557.609 66.1823 557.609 67.0469C557.609 67.8281 557.849 68.4115 558.328 68.7969C558.807 69.1823 560 69.5573 561.906 69.9219L563.062 70.1406C564.49 70.4115 565.568 70.9531 566.297 71.7656C567.036 72.5781 567.406 73.6354 567.406 74.9375C567.406 76.6667 566.792 78.0208 565.562 79C564.333 79.9688 562.625 80.4531 560.438 80.4531C559.573 80.4531 558.667 80.3594 557.719 80.1719C556.771 79.9948 555.745 79.724 554.641 79.3594V76.3906C555.714 76.9427 556.74 77.3594 557.719 77.6406C558.698 77.9115 559.625 78.0469 560.5 78.0469C561.771 78.0469 562.755 77.7917 563.453 77.2812C564.151 76.7604 564.5 76.0365 564.5 75.1094C564.5 73.776 563.224 72.8542 560.672 72.3438L560.547 72.3125L559.469 72.0938C557.812 71.7708 556.604 71.2292 555.844 70.4688C555.083 69.6979 554.703 68.651 554.703 67.3281C554.703 65.651 555.271 64.3594 556.406 63.4531C557.542 62.5365 559.161 62.0781 561.266 62.0781C562.203 62.0781 563.104 62.1667 563.969 62.3438C564.833 62.5104 565.682 62.7656 566.516 63.1094ZM585.797 63.1094V65.9219C584.974 65.4427 584.146 65.0833 583.312 64.8438C582.479 64.6042 581.63 64.4844 580.766 64.4844C579.464 64.4844 578.49 64.6979 577.844 65.125C577.208 65.5417 576.891 66.1823 576.891 67.0469C576.891 67.8281 577.13 68.4115 577.609 68.7969C578.089 69.1823 579.281 69.5573 581.188 69.9219L582.344 70.1406C583.771 70.4115 584.849 70.9531 585.578 71.7656C586.318 72.5781 586.688 73.6354 586.688 74.9375C586.688 76.6667 586.073 78.0208 584.844 79C583.615 79.9688 581.906 80.4531 579.719 80.4531C578.854 80.4531 577.948 80.3594 577 80.1719C576.052 79.9948 575.026 79.724 573.922 79.3594V76.3906C574.995 76.9427 576.021 77.3594 577 77.6406C577.979 77.9115 578.906 78.0469 579.781 78.0469C581.052 78.0469 582.036 77.7917 582.734 77.2812C583.432 76.7604 583.781 76.0365 583.781 75.1094C583.781 73.776 582.505 72.8542 579.953 72.3438L579.828 72.3125L578.75 72.0938C577.094 71.7708 575.885 71.2292 575.125 70.4688C574.365 69.6979 573.984 68.651 573.984 67.3281C573.984 65.651 574.552 64.3594 575.688 63.4531C576.823 62.5365 578.443 62.0781 580.547 62.0781C581.484 62.0781 582.385 62.1667 583.25 62.3438C584.115 62.5104 584.964 62.7656 585.797 63.1094ZM591.25 72.1875H607.766V74.8281H591.25V72.1875ZM591.25 64.8281H607.766V67.4688H591.25V64.8281ZM623.922 55.8281V65.3906H620.422V55.8281H623.922ZM615.594 55.8281V65.3906H612.094V55.8281H615.594ZM644.859 69.1562V80H641.969V69.1562C641.969 67.5833 641.693 66.4271 641.141 65.6875C640.589 64.9479 639.724 64.5781 638.547 64.5781C637.203 64.5781 636.167 65.0573 635.438 66.0156C634.719 66.9635 634.359 68.3281 634.359 70.1094V80H631.484V55.6875H634.359V65.125C634.87 64.125 635.562 63.3698 636.438 62.8594C637.312 62.3385 638.349 62.0781 639.547 62.0781C641.328 62.0781 642.656 62.6667 643.531 63.8438C644.417 65.0104 644.859 66.7812 644.859 69.1562ZM665.219 69.8125V71.2188H652.766V71.3125C652.682 73.6979 653.135 75.4115 654.125 76.4531C655.125 77.4948 656.531 78.0156 658.344 78.0156C659.26 78.0156 660.219 77.8698 661.219 77.5781C662.219 77.2865 663.286 76.8438 664.422 76.25V79.1094C663.328 79.5573 662.271 79.8906 661.25 80.1094C660.24 80.3385 659.26 80.4531 658.312 80.4531C655.594 80.4531 653.469 79.6406 651.938 78.0156C650.406 76.3802 649.641 74.1302 649.641 71.2656C649.641 68.474 650.391 66.2448 651.891 64.5781C653.391 62.9115 655.391 62.0781 657.891 62.0781C660.12 62.0781 661.875 62.8333 663.156 64.3438C664.448 65.8542 665.135 67.6771 665.219 69.8125ZM662.344 68.9688C662.219 67.7604 661.776 66.7188 661.016 65.8438C660.266 64.9583 659.182 64.5156 657.766 64.5156C656.38 64.5156 655.24 64.974 654.344 65.8906C653.448 66.8073 652.958 67.8385 652.875 68.9844L662.344 68.9688ZM678.234 73.6719C678.234 74.9635 678.469 75.9375 678.938 76.5938C679.417 77.25 680.12 77.5781 681.047 77.5781H684.406V80H680.766C679.047 80 677.714 79.4531 676.766 78.3594C675.828 77.2552 675.359 75.6927 675.359 73.6719V57.8438H670.75V55.5938H678.234V73.6719ZM697.516 73.6719C697.516 74.9635 697.75 75.9375 698.219 76.5938C698.698 77.25 699.401 77.5781 700.328 77.5781H703.688V80H700.047C698.328 80 696.995 79.4531 696.047 78.3594C695.109 77.2552 694.641 75.6927 694.641 73.6719V57.8438H690.031V55.5938H697.516V73.6719ZM715.188 64.5156C713.729 64.5156 712.625 65.0833 711.875 66.2188C711.125 67.3542 710.75 69.0365 710.75 71.2656C710.75 73.4844 711.125 75.1667 711.875 76.3125C712.625 77.4479 713.729 78.0156 715.188 78.0156C716.656 78.0156 717.766 77.4479 718.516 76.3125C719.266 75.1667 719.641 73.4844 719.641 71.2656C719.641 69.0365 719.266 67.3542 718.516 66.2188C717.766 65.0833 716.656 64.5156 715.188 64.5156ZM715.188 62.0781C717.615 62.0781 719.469 62.8646 720.75 64.4375C722.042 66.0104 722.688 68.2865 722.688 71.2656C722.688 74.2552 722.047 76.5365 720.766 78.1094C719.484 79.6719 717.625 80.4531 715.188 80.4531C712.76 80.4531 710.906 79.6719 709.625 78.1094C708.344 76.5365 707.703 74.2552 707.703 71.2656C707.703 68.2865 708.344 66.0104 709.625 64.4375C710.906 62.8646 712.76 62.0781 715.188 62.0781ZM739.609 55.8281V65.3906H736.109V55.8281H739.609ZM731.281 55.8281V65.3906H727.781V55.8281H731.281Z"
-				fill="#e9ae7e"
-			/>
-			<path
-				d="M765.547 56.6719H768.719V65.6094H777.359V56.6719H780.531V80H777.359V68.2656H768.719V80H765.547V56.6719ZM800.188 69.8125V71.2188H787.734V71.3125C787.651 73.6979 788.104 75.4115 789.094 76.4531C790.094 77.4948 791.5 78.0156 793.312 78.0156C794.229 78.0156 795.188 77.8698 796.188 77.5781C797.188 77.2865 798.255 76.8438 799.391 76.25V79.1094C798.297 79.5573 797.24 79.8906 796.219 80.1094C795.208 80.3385 794.229 80.4531 793.281 80.4531C790.562 80.4531 788.438 79.6406 786.906 78.0156C785.375 76.3802 784.609 74.1302 784.609 71.2656C784.609 68.474 785.359 66.2448 786.859 64.5781C788.359 62.9115 790.359 62.0781 792.859 62.0781C795.089 62.0781 796.844 62.8333 798.125 64.3438C799.417 65.8542 800.104 67.6771 800.188 69.8125ZM797.312 68.9688C797.188 67.7604 796.745 66.7188 795.984 65.8438C795.234 64.9583 794.151 64.5156 792.734 64.5156C791.349 64.5156 790.208 64.974 789.312 65.8906C788.417 66.8073 787.927 67.8385 787.844 68.9844L797.312 68.9688ZM813.203 73.6719C813.203 74.9635 813.438 75.9375 813.906 76.5938C814.385 77.25 815.089 77.5781 816.016 77.5781H819.375V80H815.734C814.016 80 812.682 79.4531 811.734 78.3594C810.797 77.2552 810.328 75.6927 810.328 73.6719V57.8438H805.719V55.5938H813.203V73.6719ZM832.484 73.6719C832.484 74.9635 832.719 75.9375 833.188 76.5938C833.667 77.25 834.37 77.5781 835.297 77.5781H838.656V80H835.016C833.297 80 831.964 79.4531 831.016 78.3594C830.078 77.2552 829.609 75.6927 829.609 73.6719V57.8438H825V55.5938H832.484V73.6719ZM850.156 64.5156C848.698 64.5156 847.594 65.0833 846.844 66.2188C846.094 67.3542 845.719 69.0365 845.719 71.2656C845.719 73.4844 846.094 75.1667 846.844 76.3125C847.594 77.4479 848.698 78.0156 850.156 78.0156C851.625 78.0156 852.734 77.4479 853.484 76.3125C854.234 75.1667 854.609 73.4844 854.609 71.2656C854.609 69.0365 854.234 67.3542 853.484 66.2188C852.734 65.0833 851.625 64.5156 850.156 64.5156ZM850.156 62.0781C852.583 62.0781 854.438 62.8646 855.719 64.4375C857.01 66.0104 857.656 68.2865 857.656 71.2656C857.656 74.2552 857.016 76.5365 855.734 78.1094C854.453 79.6719 852.594 80.4531 850.156 80.4531C847.729 80.4531 845.875 79.6719 844.594 78.1094C843.312 76.5365 842.672 74.2552 842.672 71.2656C842.672 68.2865 843.312 66.0104 844.594 64.4375C845.875 62.8646 847.729 62.0781 850.156 62.0781ZM866.719 74.3594H871.594V78.5L868.125 85.2344L865 85.25L866.719 78.5V74.3594ZM898.375 62.5H901.219L904.266 76.6406L906.766 67.6094H909.219L911.75 76.6406L914.797 62.5H917.641L913.547 80H910.797L908 70.4062L905.219 80H902.469L898.375 62.5ZM927.281 64.5156C925.823 64.5156 924.719 65.0833 923.969 66.2188C923.219 67.3542 922.844 69.0365 922.844 71.2656C922.844 73.4844 923.219 75.1667 923.969 76.3125C924.719 77.4479 925.823 78.0156 927.281 78.0156C928.75 78.0156 929.859 77.4479 930.609 76.3125C931.359 75.1667 931.734 73.4844 931.734 71.2656C931.734 69.0365 931.359 67.3542 930.609 66.2188C929.859 65.0833 928.75 64.5156 927.281 64.5156ZM927.281 62.0781C929.708 62.0781 931.562 62.8646 932.844 64.4375C934.135 66.0104 934.781 68.2865 934.781 71.2656C934.781 74.2552 934.141 76.5365 932.859 78.1094C931.578 79.6719 929.719 80.4531 927.281 80.4531C924.854 80.4531 923 79.6719 921.719 78.1094C920.438 76.5365 919.797 74.2552 919.797 71.2656C919.797 68.2865 920.438 66.0104 921.719 64.4375C923 62.8646 924.854 62.0781 927.281 62.0781ZM954.984 66.1094C954.37 65.6302 953.745 65.2812 953.109 65.0625C952.474 64.8438 951.776 64.7344 951.016 64.7344C949.224 64.7344 947.854 65.2969 946.906 66.4219C945.958 67.5469 945.484 69.1719 945.484 71.2969V80H942.594V62.5H945.484V65.9219C945.964 64.6823 946.698 63.7344 947.688 63.0781C948.688 62.4115 949.87 62.0781 951.234 62.0781C951.943 62.0781 952.604 62.1667 953.219 62.3438C953.833 62.5208 954.422 62.7969 954.984 63.1719V66.1094ZM967.453 73.6719C967.453 74.9635 967.688 75.9375 968.156 76.5938C968.635 77.25 969.339 77.5781 970.266 77.5781H973.625V80H969.984C968.266 80 966.932 79.4531 965.984 78.3594C965.047 77.2552 964.578 75.6927 964.578 73.6719V57.8438H959.969V55.5938H967.453V73.6719ZM988.906 64.7344V55.6875H991.781V80H988.906V77.7969C988.427 78.6615 987.786 79.3229 986.984 79.7812C986.193 80.2292 985.276 80.4531 984.234 80.4531C982.12 80.4531 980.453 79.6354 979.234 78C978.026 76.3542 977.422 74.0885 977.422 71.2031C977.422 68.3594 978.031 66.1302 979.25 64.5156C980.469 62.8906 982.13 62.0781 984.234 62.0781C985.286 62.0781 986.214 62.3073 987.016 62.7656C987.818 63.2135 988.448 63.8698 988.906 64.7344ZM980.453 71.2656C980.453 73.4948 980.807 75.1771 981.516 76.3125C982.224 77.4479 983.271 78.0156 984.656 78.0156C986.042 78.0156 987.094 77.4427 987.812 76.2969C988.542 75.151 988.906 73.474 988.906 71.2656C988.906 69.0469 988.542 67.3698 987.812 66.2344C987.094 65.0885 986.042 64.5156 984.656 64.5156C983.271 64.5156 982.224 65.0833 981.516 66.2188C980.807 67.3542 980.453 69.0365 980.453 71.2656ZM1003.31 56.6719H1006.48V66.9062L1006.16 71.7031H1003.64L1003.31 66.9062V56.6719ZM1002.84 76.0312H1006.95V80H1002.84V76.0312Z"
-				fill="#c699e3"
-			/>
-		</svg>
-	`;
-}
